billboard-streamlit/requirements.txt
streamlit==1.41.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
pandas==2.2.3

pytest==8.3.4
ruff==0.8.4
